--- menpo/groupalign/__init__.py
@@ -1,1 +1,0 @@
-from menpo.groupalign.procrustes import GeneralizedProcrustesAnalysis
