+++ menpo/transform/__init__.py
@@ -3,3 +3,5 @@ from .homogeneous import *
 from .thinplatesplines import ThinPlateSplines
 from .piecewiseaffine import PiecewiseAffine
 from .rbf import R2LogR2RBF, R2LogRRBF
+
+from .groupalign.procrustes import GeneralizedProcrustesAnalysis
